Committed changes to make the code compile on Solaris. This patch is losely based on the discussion in   http://code.google.com/p/google-authenticator/issues/detail?id=76 Thanks to "codermonkey", and "Steve <kb7sqi>" for figuring out how Solaris differs from the other environments that we have tested this code in.
I am not happy about having to use "-mimpure-text", but for lack of a Solaris
machine, I cannot debug why the existing "-fPIC" options are insufficient.
Feel free to suggest an alternative option, if you are aware of it.

Also, everybody who was affected by this bug, please check that these changes
in fact resolve all issues on Solaris.