Fixed duplicate prompt bug in libpam module.
When the empty string is returned after the prompt, the module no longer prompts
a second time.  (That second prompt was ignored anyway, even if it was correct,
in my testing.)  That returned empty string was leaked as well.

Also added extensive unit testing to verify that the conversation callback is
not being executed more times than expected.