Add support for issuer in google-authenticator otpauth:// URL.
As per issue #391, allow to specify a &issuer=<issuer> in otpauth:// generated
URL done by google-authenticator. By default it'll use the hostname. Setting it
to an empty value like --issuer='' prevents &issuer= to be appended to the URL
all together.

This is in line with the wiki page on
https://github.com/google/google-authenticator/wiki/KeyUriFormat
recommending to use both an issuer label prefix and an issuer parameter.