Improve PAM API conformance by using pam_get_user.
The pam_start(3) manual page (on Debian 8) says:

"The user argument can specify the name of the target user and will be
stored as PAM_USER item. If the argument is NULL, the module has to ask
for this item if necessary."

The manual page for pam_get_item(3) (current mechanism which the module
uses to obtain the user name) says:

"If a service module wishes to obtain the name of the user, it should
not use this function, but instead perform a call to pam_get_user(3)."

This patch will fix the situation whereby PAM_USER isn't set at the
time that the module is invoked (e.g. because it is the first PAM
module in the stack on the system in question, and the application
which is attempting authentication via PAM has (legitimately) not set
the user in it's call to pam_start().

Current behaviour breaks when pam_google_authenticator is used with
openvpn's openvpn-plugin-auth-pam.so, and also xrdp:

https://github.com/google/google-authenticator/issues/361

Signed-off-by: Tim Small <tim@seoss.co.uk>