Reduce secret file restrictions
This commit implements 2 options for the pam_google_authenticator
module.

`allow_perm=<perm_octal>` provides a second set of permissions to allow
on the secrets file. The default permissions required are 0400.

`no_strict_owner` allows the file to be owned by a user or group other
than the user being authenticated (or the user the process is running
as).