Add `no_strict_owner` option.
DANGEROUS OPTION!

By default the PAM module requires that the secrets file must be owned
the user logging in (or if `user=` is specified, owned by that
user). This option disables that check.

This option can be used to allow daemons not running as root to still
handle configuration files not owned by that user, for example owned by
the users themselves.

This commit is pulled out of
https://github.com/google/google-authenticator/pull/542
by Philip Woolford <woolford.philip@gmail.com>